Revert "Support custom model (#49)"
This reverts commit ec6f8db055dd7c06851238955f9fab71a9aefc39.
