+++ cosmos_rl/policy/model/deepseek_v3/parallelize.py
@@ -33,9 +33,9 @@ from torch.distributed.tensor.parallel import (
     RowwiseParallel,
     SequenceParallel,
 )
-import cosmos_rl.utils.util as util
 from cosmos_rl.utils.parallelism import ParallelDims
 from cosmos_rl.utils.logging import logger
+from cosmos_rl.utils.util import str2torch_dtype
 from cosmos_rl.policy.config import Config as CosmosConfig
 from cosmos_rl.patch import PipelineStage, Schedule1F1B, ScheduleGPipe
 from typing import Callable, Optional
@@ -87,8 +87,8 @@ def parallelize(
         apply_fsdp(
             model,
             world_mesh[tuple(dp_mesh_dim_names)],
-            param_dtype=util.str2torch_dtype(config.train.param_dtype),
-            reduce_dtype=util.str2torch_dtype(config.train.fsdp_reduce_dtype),
+            param_dtype=str2torch_dtype(config.train.param_dtype),
+            reduce_dtype=str2torch_dtype(config.train.fsdp_reduce_dtype),
             pp_enabled=parallel_dims.pp_enabled,
             cpu_offload=config.train.fsdp_offload,
             reshard_after_forward_policy=config.train.fsdp_reshard_after_forward,
