+++ cosmos_rl/policy/model/qwen2_5_vl/__init__.py
@@ -33,13 +33,9 @@ from safetensors import safe_open
 from cosmos_rl.policy.model.qwen2_5_vl.weight_converter import (
     convert_weight_from_hf,
 )
-from cosmos_rl.dispatcher.data.packer.qwen2_5_vlm_data_packer import (
-    Qwen2_5_VLM_DataPacker,
-)
-from cosmos_rl.policy.model.qwen2_5_vl.weight_mapper import QwenVL25WeightMapper
 from cosmos_rl.utils.parallelism import ParallelDims
 from cosmos_rl.policy.config import Config as CosmosConfig
-from cosmos_rl.policy.model.base import ModelRegistry, BaseModel
+from cosmos_rl.policy.model.base import BaseModel
 from functools import cached_property
 import re
 from functools import partial
@@ -848,7 +844,7 @@ class Qwen2_5_VLModel(nn.Module):
         )
 
 
-@ModelRegistry.register(Qwen2_5_VLM_DataPacker, QwenVL25WeightMapper)
+@BaseModel.register()
 class Qwen2_5_VLConditionalModel(BaseModel):
     def __init__(self, config: Qwen2_5_VL_LM_Args):
         super().__init__(config.hf_config)
