+++ cosmos_rl/rollout/weight_mapper/gpt.py
@@ -16,7 +16,7 @@
 from vllm.model_executor.models.qwen2 import Qwen2ForCausalLM
 import torch
 from typing import List, Tuple, Dict
-from cosmos_rl.policy.model.base import WeightMapper
+from cosmos_rl.rollout.weight_mapper.base import WeightMapper
 from cosmos_rl.utils.parallelism import ParallelismConfig
 from cosmos_rl.utils.parallelism_registry import (
     get_policy_parallelism_strategy,
@@ -24,8 +24,10 @@ from cosmos_rl.utils.parallelism_registry import (
 )
 from cosmos_rl.utils import util
 from transformers import AutoConfig
+from cosmos_rl.policy.model.gpt import GPT
 
 
+@WeightMapper.register_class(GPT.supported_model_types())
 class GPTWeightMapper(WeightMapper):
     def __init__(self, hf_config: AutoConfig):
         super().__init__(hf_config)
