+++ tools/dataset/gsm8k_grpo.py
@@ -199,5 +199,6 @@ if __name__ == "__main__":
         # Optional: if not provided, the default data packer of the selected model will be used
         data_packer=GSM8kDataPacker(),
         val_dataset=val_dataset,
+        val_reward_fns=[custom_reward_fn],
         val_data_packer=GSM8kDataPacker(),
     )
